Fix: Navigating from user profile to membership, throws console error (#8623)
* Fix: Navigating from user profile to membership, throws console error

* address CodeRabbit comments

* A11y: 4.1 (#8614)

* a11y 4.1.2 errors and missing keys,

* fix 412 a11y issues

* get rid of some more a11y 412 errors

* do not have nested buttons;
onClick on 'contributionProps' should not be the reason the whole card becomes a button

* add aria-label

* fixes after bot comments

* add alt to avatars

* bot suggestion fixes

* fixes

* more fixes

* last fixes

* alt whole name

* cleanup naming and regressions

* last fixes

* address comments

* Add dependencies to useMemo

---------


* address CR comment

---------
